requirements.txt
flask
notion-client
python-dotenv
requests
Flask-Limiter
google-genai
mistune
tenacity
python-docx
pdfplumber
werkzeug
